% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenCSV-package.R
\docType{package}
\name{brokenCSV}
\alias{brokenCSV}
\alias{brokenCSV-package}
\alias{package-brokenCSV}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

