% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comma_broken_csv.R
\name{comma_broken_csv}
\alias{comma_broken_csv}
\title{Generates Report on Comma-Broken Status of CSV(s)}
\usage{
comma_broken_csv(path = ".")
}
\arguments{
\item{path}{path-to-file or path-to-folder}
}
\value{
print report to console (and if multiple, \code{sink}s file to parent folder and opens report with \code{file.show()})
}
\description{
Generates Report on Comma-Broken Status of CSV(s), including broken rows and potential offending columns needing to be quoted in file.
}
\details{
If path param is a folder, for readability a report named \file{csv_comma_broken_report.txt} is dropped into path folder and opened automatically to source pane. But if path param is only a file, then the report is only printed to console.
}
\examples{
\dontrun{
comma_broken_csv(getwd())
# or
comma_broken_csv("path_to_file")

# Or if you want to run a test on multiple files
# not contained in the same (grand)parent folder,
# then lapply over the file paths like so:
lapply(c("file.path1", "file.path2"), comma_broken_csv))
}
}

